transcendence/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.60.0
django-bootstrap-v5==1.0.11
django-widget-tweaks==1.5.0
idna==3.6
nodejs==0.1.1
npm==0.1.1
npx==0.1.1
numpy==1.26.2
oauthlib==3.2.2
optional-django==0.1.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
channels~=4.0.0